backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
duckdb==0.9.2
httpx==0.25.2
apscheduler==3.10.4
pydantic==2.5.0
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
ruff==0.1.6
pytest==7.4.3
pytest-asyncio==0.21.1